#29 #30 1.introduced MapStruct mapper for mapping only necessary fields when customer is querying all items. 2.create pagination api for customer to query all items by pagination and sorted by some field.
