+++ backend/src/main/java/com/amazon/controller/CustomerItemController.java
@@ -1,8 +1,11 @@
 package com.amazon.controller;
 
+import com.amazon.dto.ItemListDto;
 import com.amazon.entity.Item;
+import com.amazon.mapper.ItemListMapper;
 import com.amazon.service.CustomerItemService;
 import org.springframework.data.domain.Page;
+import org.springframework.data.domain.PageImpl;
 import org.springframework.data.domain.Pageable;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -22,8 +25,10 @@ public class CustomerItemController {
     }
 
     @GetMapping
-    public Page<Item> findAllItems(Pageable pageable) {
-        return null;
+    public Page<ItemListDto> findAllItems(Pageable pageable) {
+        Page<Item> itemPage = customerItemService.findAllItems(pageable);
+        List<ItemListDto> itemListDtos = ItemListMapper.INSTANCE.map(itemPage.getContent());
+        return new PageImpl<>(itemListDtos, pageable, itemPage.getTotalElements());
     }
 
     public List<Item> findAllByname(String name){
@@ -32,8 +37,7 @@ public class CustomerItemController {
     public List<Item> findAllByBrand(String brand){
         return customerItemService.findAllByBrand(brand);
     }
-        //Page<ItemDto> 再放pageable
-        //返回给前端
+
     @PostMapping
     public Item saveItem() {
         return new Item();
