+++ backend/src/main/java/com/amazon/dto/ItemListDto.java
@@ -1,0 +1,41 @@
+package com.amazon.dto;
+
+public class ItemListDto {
+
+    private int id;
+    private String name;
+    private double price;
+    private String brand;
+
+    public int getId() {
+        return id;
+    }
+
+    public void setId(int id) {
+        this.id = id;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+
+    public double getPrice() {
+        return price;
+    }
+
+    public void setPrice(double price) {
+        this.price = price;
+    }
+
+    public String getBrand() {
+        return brand;
+    }
+
+    public void setBrand(String brand) {
+        this.brand = brand;
+    }
+}
