+++ backend/src/main/java/com/amazon/service/CustomerItemService.java
@@ -3,10 +3,12 @@ package com.amazon.service;
 import com.amazon.entity.Item;
 import com.amazon.repository.CustomerItemRepository;
 import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 
 import java.util.Collection;
 import java.util.List;
+import java.util.Optional;
 
 @Service
 public class CustomerItemService {
@@ -16,15 +18,17 @@ public class CustomerItemService {
     public CustomerItemService(CustomerItemRepository customerItemRepository) {
         this.customerItemRepository = customerItemRepository;
     }
+
+    public Page<Item> findAllItems(Pageable pageable) {
+        return customerItemRepository.findAll(pageable);
+    }
+
     public List<Item> findAllByBrand(String brand){
         return customerItemRepository.findAllByBrand(brand);
     }
 
     public List<Item> findAllByName(String name){
         return customerItemRepository.findAllByName(name);
-    }
-    public List<Item> findAllItems() {
-        return customerItemRepository.findAll();
     }
 
     public Item save(Item item) {
